sinkv2(ticdc): rename
