+++ cdc/sinkv2/eventsink/mq/mq_sink.go
@@ -15,7 +15,6 @@ package mq
 
 import (
 	"github.com/pingcap/tiflow/cdc/model"
-	"github.com/pingcap/tiflow/cdc/sinkv2/event"
 	"github.com/pingcap/tiflow/cdc/sinkv2/eventsink"
 )
 
@@ -26,7 +25,7 @@ var _ eventsink.EventSink[*model.RowChangedEvent] = (*Sink)(nil)
 type Sink struct{}
 
 // WriteEvents writes events to the sink.
-func (s *Sink) WriteEvents(rows ...*event.RowChangeCallbackableEvent) {
+func (s *Sink) WriteEvents(rows ...*tableevent.RowChangeCallbackableEvent) {
 	// TODO implement me
 	panic("implement me")
 }
