+++ cdc/sinkv2/tableevent/event.go
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package event
+package tableevent
 
 import (
 	"github.com/pingcap/tiflow/cdc/model"
